requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiostream==0.4.5
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.1
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.0
bleach==6.0.0
bokeh==2.4.3
boto3==1.26.112
botocore==1.29.112
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-default-group-wheel==1.2.2
cloudpickle==2.0.0
coloredlogs==15.0.1
comm==0.1.3
commonmark==0.9.1
connectorx==0.3.1
contourpy==1.0.7
cycler==0.11.0
dask==2023.3.2
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
deltalake==0.8.1
distributed==2023.3.2
duckdb==0.7.1
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.95.2
fastapi-utils==0.2.1
fastjsonschema==2.16.3
filelock==3.12.2
Flask==2.2.3
flatbuffers==23.5.26
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
grpclib==0.4.3
h11==0.14.0
h2==4.1.0
HeapDict==1.0.1
hpack==4.0.0
html5lib==1.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.1.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
llm==0.3
load-dotenv==0.1.0
locket==1.0.0
lz4==4.3.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
modal-client==0.49.2289
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
nba-api==1.2
nbclassic==0.5.4
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
onnxruntime==1.15.1
openai==0.27.2
opencv-python-headless==4.7.0.72
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
pexpect==4.8.0
pgvector==0.1.8
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.2.0
plotly==5.14.0
poetry-core==1.5.2
polars==0.16.17
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.1
psutil==5.9.4
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
pydantic==1.10.7
pyfredapi==0.5.1
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.1
regex==2023.5.5
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.9.0
Send2Trash==1.8.0
sigtools==4.0.1
six==1.16.0
sklearn==0.0.post5
slack-bolt==1.18.0
slack-sdk==3.21.3
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==1.4.48
sqlite-fts4==1.0.3
sqlite-utils==3.31
stack-data==0.6.2
starlette==0.27.0
stomper==0.4.3
strip-tags==0.2
sympy==1.12
synchronicity==0.5.2
tabulate==0.9.0
tblib==1.7.0
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.1.0
tiktoken==0.4.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
-e git+https://github.com/aviral-a/pytorch.git@a89320e340b2f40d8d5bb6c2b52971458938b3f5#egg=torch
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
ttok==0.1
typer==0.9.0
types-certifi==2021.10.8.3
types-toml==0.10.8.6
typing_extensions==4.5.0
undetected-chromedriver==3.2.1
uri-template==1.2.0
urllib3==1.26.15
uvicorn==0.22.0
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.2
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wsproto==1.2.0
xlsx2csv==0.8.1
XlsxWriter==3.0.9
yarl==1.8.2
zict==2.2.0
zipp==3.15.0
